Fix error in HSU_DataFullFolderPathString
The old implementation assumed that panelTitle refers to an existing
window. This popped up at start of the databrowser.

Now we add a fallback which just parses the panelTitle if it does not exist.
For that purpose the generic functions
ParseDeviceString
and
BuildDeviceString
have been introduced.

Rename Path_ListOfYokedDACs(...) to GUIListOfYokedDACs() as it contains fallback
code for an empty list. The generic part of Path_ListOfYokedDACs is now contained in
GetListOfYokedDACs().

Original-CommitID: 3d308998a49a484f9c19a76b2bbb177b6f4658bd