Ignore errors on closing the device
This helps in the common situation that you saved
an experiment with locked ITC devices.
Now the ITC XOP will not complain about a unknown device ID
on pressing "Unlock Device" with no device being locked.

A note about the "/Z flag" for the ITC* XOP calls.
This is a true or false flag.

One of the following methods set it to true
/Z
/Z=1
/Z=Something
and only this invocation to false.
/Z

Original-CommitID: c317d4f96452b86d11eb939ec39008c84bfcf989