Fix line count and preserve leading whitespace
The doxygen documentation of the INPUT_FILTER configuration variable
states:
 # Note that the filter must not add or remove lines; it is applied before the
 # code is scanned, but not when the output code is generated. If lines are added
 # or removed, the anchors will not be placed correctly.

There are now no lines removed and the only line added is at the end of the file.
And that does not matter.

Additionally any leading whitespace is readded in the output in order to
not mess up the indentation.

Original-CommitID: 8e2eca205fe9030de7d51e230c27308448d244b8