Bugfix: Initialize the settingsHistory wave's layer count to NUM_HEADSTAGES
This prevents an runtime error if "Save Amp settings" is enabled and the labnotebook does not exist.
In that case the resizing of the layerCount for keyColCount == 2 will resize it to zero as the first incoming wave
is the amplifier settings wave. And then later settings waves with layer counts larger than zero can not be added.

Original-CommitID: 6148630aee40a8fa42a5ee5ad2f236f51553bfb7