Limit the duration of an epoch to 30 minutes
This avoids errors like out of memory and too many points
in the wave from Igor itself.

As we now might abort in WB_GetSegmentWave we also have to change the WP
copying logic so that we don't need to the resetting WP step.

Original-CommitID: a2bf8da0d3cbcfda70c113779f39009ae0695464