Bugfix: Correct search for exising sweeps in graphs
Currently we ignore sweeps that are already shown in the sweep browser.
The existing implementation was flawed as it did not check that the
sweep properties (expName, expFolder, device, sweep) are found in the
same row in the map. And not just somewhere.

Original-CommitID: 59e56524a4a13daf737725f0827c5b378f6f49e7