Regression bugfix: Ramp creation in the Wavebuilder
Since 15c7fb36 (Move all segment creation parameters into a single structure, 2015-04-15)
the amplitude of the created ramp is incorrect.

The reason for the bug is that the following conversion
a/(b/c) -> a/b/c
used in 15c7fb36 is not true.

Instead a valid conversion is
a/(b/c) -> a*c/b
which is now used.

Original-CommitID: 410583abdbe340f18b3d1b90fcb9ffe9e1d71920