Add a notebook which helps the user track sweep comments
Next to the comment string in the DA_Ephys panel a button now allows to
create a notebook which shows all added comments. The comments are
manually added when the "NB" button is pressed.

On killing the notebook, experiment save or unlocking the panel the
contents of the notebook are serialized in a global string variable,
located in GetUserComment(...). To avoid confusion the notebook is write
protected if the DA_Ephys panel is unlocked.

Additionally comments by users into the notebook are always preserved.

On "Save and clear" the notebook contents are cleared.

The comments are now also only written to the labnotebook on pressing the
NB button.

Original-CommitID: cbb3f4cbfd79f5e38392e519287330e10fcc7da7