Remove duplicated check-for-save-flag pattern
At all call sites we have the same pattern. Move the code into
DM_SaveITCData and rename the function to DM_SaveAndScaleITCData.

The function DM_SaveITCData is superfluous as it is always called for
ITCDataWave. Instead of copying ITCDataWave and performing the AD scaling
on the original and the copy we just scale the original now.

Original-CommitID: e83966a3999f31ff408d17d9e15c8ff1e7d56c11