Wavebuilder: Write the total number of epochs and the step count directly into SegWvType
This allows to query the values from the wave later on instead of having
to query the GUI.

Also remove obsolete parts from WBP_SaveSetParam as the values are already
written into SegWvType.

This makes it necessary to write the initial values for the number of
steps and epochs to SegWvType on creation.

Original-CommitID: 5a587a6c7db4e3696efa337f22413fc82411a547