WaveBuilder: Fix on demand wave creation for custom wave epochs
Only newly created custom waves epochs can be recreated on demand.
The reason is that the path to the custom wave was not stored.
We now store the path and implement a heuristic for determining if we have
new or old style entries.

Original-CommitID: 11aad4b802444956758d00a57478be5e7fe14797