Bugfix: Always set a minimum onset delay when using test pulse insertion
In the case that the existing onset delay is big enough to hold the
testpulse the lower limit of the control was not set.
This allowed the user to set a lower onset delay later on, thus possibly
creating an onset delay which is too small for the testpulse.

Original-CommitID: 36131ad5de633d7d5f1084bd17f66eec7cc1a7d8