AI_UpdateAmpModel: Fix setting values for non-current headstages
The old implementation always overwrites the passed headstage if
"send to all amps" is turned off.

We now don't do that anymore, instead we allow to pass an value in
addition so that we don't have to query the GUI for that element.
If we don't query the GUI we at least try to sync it via calling
AI_UpdateAmpView.

Convert also multiple call sites to use the new interface.

Original-CommitID: d58171bf2e83b98aa8d0b3348b7014fb0fc22878