DAP_StopOngoingDataAcquisition: Only save the sweep data if sensible
For repeated background acqusition the user can abort the DAQ at any time.
If the abort happens during the ITI, with or without Testpulse, the sweep data
was saved. Although we know that it was completly zero.

Now we only save it if neither Testpulse nor ITI is running.

Original-CommitID: 51fdf6b84fb43cd367d2f43ab47123207a9bc48b