Store the current testpulse running mode during acquisition
The mode is stored in $GetTestpulseRunMode(panelTitle) and allows to
stop the testpulse knowing if it is multi/single device.

This fixes a potential bug as just checking that the background operation
is running might be inaccurate if multiple devices are active.

This also removes the dependency of P_IsTPActive, now called
TP_CheckIfTestpulseIsRunning, on the disabled/enabled state of the
testpulse button.

Original-CommitID: addf2f3cc9703d45891ea1806daee5d495dd944e