Remove DA/TTL channel control adaptations for the testpulse
Before a testpulse could be started we always did multiple changes to the
DA/TTL GUI controls:
- Store the currently selected stimsets
- Select the testpulse stimsets
- Set the DAScale to 1 (VOLTAGE_CLAMP) or amplitudeIC/amplitudeVC
- Store the currently selected TTL stimsets

All these changes were reverted once the testpulse finished execution
without errors.
In case of errors the user had to manually change the settings back.

This was quite error prone and in order to avoid that we now don't do
anything special with the controls anymore.

The testpulse stimset is now set to 1 for the peak range and the scaling for the IC/VC
amplitude is done in DC_PlaceDataInITCDataWave.

SI_CalculateMinSampInterval learned to ignore TTL channels if calculating
the sampling interval for the testpulse.

DC_LongestOutputWave and DC_PlaceDataInITCDataWave now just uses
GetTestPulse() as stimset for the testpulse.

Removed Functions:
- DAP_StoreTTLState, DAP_RestoreTTLState
- TP_StoreSelectedDACWaves, TP_ResetSelectedDACWaves
- TP_StoreDAScale, TP_SetDAScaleToOne, TP_RestoreDAScale

The testpulse stimset is *not* removed from the DA popup menues as the
index of the stimsets might be used in other software further downstream.

Original-CommitID: a0c98f790112eecfb76e1ad0ceed04c888ac3674