NWB: Rework session_start_time handling
The previous definition of the session_start_time was basically identical
to the file_create_date.

Now we set the session_start_time as the time of the locking of the last
device. When creating a NWB file this time is then used as
sesstion_starting_time. If we reopen an existing NWB file, we read the
time back from the NWB file.

Original-CommitID: 12e03b10725e95930db5e29d874af1766933d562