DAP_CheckSettings: Find duplicate headstage <-> channel associations
For functions like AFH_GetHeadstageFromADC it is crucial that the
headstage AD/DA channel mapping is one to one.

Until now we have never enforced this. The symptom of a violation of this
precondition is that sweep dependent data for the labnotebook might end up
in the wrong layer, i.e. is connected with the wrong headstage.

Check for these before for DAQ and TP.

Original-CommitID: 044ede098f0131b40f620984170762f64d54d878