Fix help messages for controls inside TabControls
As it turns out the help messages of the controls are currently not shown
if the TabControl is defined later (or further down in a panel macro) than
its control inside. Instead the help messages of the TabControl are shown.

Move the TabControls therefore in front of the containing controls. And
also remove the tab help message to avoid confusion.

This will not be necessary for IP7 versions starting from tomorrow's
nightly build.

Original-CommitID: 0bfd39158b04e474082f01f78ba037b0f4fee69b