Display the scaled AD data also during DAQ
Unscaled AD channels are a great confusion for new users.
Therefore we now employ a similiar strategy as for the testpulse.

This requires to copy the contents of ITCDataWave to OscilloscopeData on
each change of the opportunity where the data might have changed.

Unnecessary "hand made" updates to the ITCDataWave are due to this change
also not needed anymore.

Original-CommitID: 77a74262995b5527e17590849fb572ea04facfd7