WaveBuilderPanel: Fix duration control disabling/enabling logic
Since 68bc2b7c (Allow the user to define the number of pulses instead of
the duration for the square train pulse segment, 2015 -04-15) we have
support for defining the Square Pulse Train epoch in terms of number of
pulses instead of duration. The therein implemented logic for disabling the
duration control failed to reenable it on epoch/tab change.

Original-CommitID: 08268e8fb11a490f5bda3291bedeff3e8dd7ea6d