Fix functions creating list of paths
As it turns out the semicolon (";") is a valid character on Windows [1].
Therefore functions creating lists of files can not use the semicolon
as list separator.

Instead we now use the pipe symbol ("|").
Additionally we remove the duplicated implementation
GetFilesRecursively. This functions predates
GetAllFilesRecursivelyFromPath but did not handle shortcuts.

This bug is also present in the internal WaveMetrics implementation of
IndexedDir and IndexedFile as both return a semiclon separated list if
passed -1 as index.

[1]: https://msdn.microsoft.com/en-us/library/aa365247.aspx

Original-CommitID: 60d4028b9491f4f5f7f5306e49d12206fc148922