Multi device TP: Prevent endless device restarting loop
If we restarted the device the fifo position in
ITCFIFOAvailAllConfigWave[ADChannelToMonitor][2] is zero and
ActiveDeviceList[i][5] is zero as well.

If we are then executed fast enough we might still have a zero fifo
position. But as that is no problem per se we must prevent the device
restarting in that case.

Original-CommitID: 24cc9df28f4a90e3b91f18d45df4b15cc1e525f3