Oscilloscope: Rework graph updating logic
The current approach has sometimes problems with updating the axes ranges.
This is especially true for multi device TP.

Rework that logic and avoid the live keyword. We now use autoscaling to
get the axis range and then set that axis range, if it is a valid range,
manually. This gives us a non-autoscaled axis and updates the axis ranges
from time to time.

Original-CommitID: 9682a1608d68aaddb49f5904ce86a68344a78595