Yoked devices: Fix igor error about missing windows
The order of the arguments for AddListItem is itemStr and then listStr.
And in our case panelList is the itemStr. And then we also have to add
paneLList, which is the lead panel to the front of the list.

Bug in DAP_UpdateSweepLimitsAndDisplay introduced with the initial
implementation in dbe20bd8 (Synchronize sweep and next sweep from leader
to follower, 2014-10-20).

Bug in DAP_SyncGuiFromLeaderToFollower introduced with the initial
implementation in 2b324acc (Synchronize some more controls between leader
and follower, 2014-10-20).

Original-CommitID: 27fcb90c31cc113115e1f0fcbba9a43cfde31434