DM_UpdateOscilloscopeData: Only update with valid data
Due to timing issue with the ITC device we might call
DM_UpdateOscilloscopeData with a fifo position of zero.
In that case it is very very likely that we did not acquire any data yet.

To prevent the oscilloscope display to show us a zero for the very first
point, which will be later overwritten anyway, we do nothing if we
encounter a fifo position of zero.

Original-CommitID: ab6ccaf069292ffe8602efe5878195b388be4c34