Multi device TP: Prevent error out on yoked device set column calculation
Since cc78dbb8 (Introduce a getter function for the per device global
variable count, 2015-05-14) we use getter functions for count and rely on
the fact that non-existence of this variable and it being NaN have the
same semantics.

Unfortunately in DC_CalculateChannelColumnNo this was not done. The result
was that with yoked devices the follower panel's count variable was
already created and correctly set to NaN. But we only checked for its
existence and not its finiteness. Therefore the returned setColumn was
NaN. This only cropped up much later as MultiThread wave assignments don't
error out immediately. Just later at some point. The result was a hard to
debug error.

Now the last two places where we check for the existence of count are
replaced with checks for finiteness. In addition the count getter
functions are used.

Original-CommitID: 9ad1a66e7f4a2ec3e7061d4ec784386310bc5585