NWB: Fix exporting new sweeps with data from an existing experiment
The start time of the session is currently the time of the last the device
locking. If you now open an existing experiment with some sweeps and lock
the device and try to export into NWB you get an assertion because the
starting time of the old, existing sweeps will be negative.

Now we set, if we have existing data being acquired *before* the last
device locking, the session starting time to the starting time of the
oldest sweep.

Original-CommitID: 25b4bfadd50ad4c067d2c587bab31b8bab27cca8