Multi device yoked: Fix stopping DAQ with TP button crash
We call DAP_StopOngoingDataAcqMD when a DAQ is active and a user
requested a TP. This leads to a crash for yoked devices, as
DAP_StopOngoingDataAcq does not handle yoking.

Call the correct function ITC_StopOngoingDAQMultiDevice (renamed from
DAM_StopDAQMultiDevice) and move the functions so that we are able
to use static specifiers to prevent similiar problems in the future.

Original-CommitID: 62f68fe90cc3d3bf7f6991032056f2bdd38f4c5f