Add MIES Background Task Watcher panel
For debugging DAQ/TP related bugs it is sometimes beneficial to know
exactly what background task is running and what not. This panel here
allows that as it shows the state of each of the 6 background tasks
currently in use. Implemented in an independent module to be as robust as
possible.

The includes are inside a doxygen ignore block as doxygen treats include
directives inside namespaces differently than normal includes and
therefore complains.

Original-CommitID: 1df0b897d86a5bf80ff2b37097e6d0a9b7b838a2