DAP_SliderProc_MIESHeadStage: Avoid executing code on control kill
sc.eventCode is -1 if the control is killed, and -1 & 0x1 is true, so
the block which was supposed to be only executed on value change is also
executed on control kill.

Use our default approach using a switch statement and comparison for
equality.

Bug is present since forever.

Original-CommitID: cfba6d5f9fdda9fdae93c8df05b4ccbd2ecc9218