DAQ: Don't check the settings for stopping DAQ
Since e3d65245 (More thorough overall and headstage settings check,
2014-09-30) we check the panel settings before starting and stopping DAQ.

But stopping DAQ must not require valid settings, as e.g. in cases for
indexing these settings are not valid, and thus the user is prevented to
stop DAQ.

Now we only check the settings if we are starting DAQ. This makes it
necessary to check early on if the panel is locked.

Original-CommitID: cccb1586c214b26a01f41ff1a4eff29549a1c777