DC_PlaceDataInITCDataWave: Fix off-by-one error for Multi device TP/dDAQ
In 4d15479b (Remove DC_GlobaleChangesToITCDataWave and make code a bit
more readable, 2015-05-04) the name of the endRow variable was changed to
setLength.  But unfortunately the naming was choosen very badly as
setLength was *not* denoting a length but the last valid row index of the
stimset with decimation taken into account.

In a6a5203b (Bugfix: Change Testpulse to ITCDataWave conversion logic,
2015-07-29) this setLength was introduced for the multi device TP
calculation, but used as a length.

The result of this bug is that the testpulse edges move horizontally
during TP.

Convert setLength to be a length and convert all users using it correctly.
The users using it incorrectly, TP and dDAQ, are then therefore fixed.

Original-CommitID: 52001ad8c6269f68e7e86aaa3f2ab5a6e9970c87