Override ITI checkbox was converted to Use stim set ITI checkbox.
When only a single DA channel is active, ITI is calculation is handled as if indexing is locked.
Check box is now on the data acq tab of the DA_ephys panel next to the ITI setvar.

Now the user can fully override by not checking the get set ITI checkbox
OR they can override for a active set, but, when a new set is indexed the
corresponding set ITI will be used if the Get Set ITI checkbox is checked.

Original-CommitID: 6f79c4583ec75a23093dc505c736e0164a22a4db