Add support for I=0 clamp mode
The I=0 clamp mode is a specialization of the I clamp mode. Therefore we use
the same values for this mode in our own storage.

In this clamp mode no stimulus is reaching the cell, to make that more
obvious to the user we use a DAScale of zero in that case.

Original-CommitID: 5517aa95d21a84dad292312830857b0797aa559a