SCOPE_UpdateGraph: Avoid sequential AD channel update
The approach using SetAxis/A and GetAxis works but results in a
sequential update of all AD channels.

This is not nice from an experimentators POV.

Now we use WaveStats instead to calculate the left axis range. For nicer
viewing we make the axis a bit a larger as we woule need to. Additionally
we avoid updating the axis range if the data lies within
ADDITIONAL_SPACE_AD_GRAPH times the total axis range.

Original-CommitID: 29f2f7aff089f4d615442f6bcc25b2b96acfdcbf