Databrowser/SweepBrowser: Don't show incorrect units and channel numbers for overlayed sweeps
Until now we overwrote the left axis labels in the
Databrowser/Sweepbrowser with overlay sweeps enabled.

This results in incorrect units and channel numbers if these differ for
the overlayed sweeps.

Now we correctly output a.u. as unit and ? as channel number if the new
axis label would be different from the existing one.

Original-CommitID: 1fc8b4a810ca381570f02a50120902aaad3ba6e4