Remove AmpIndex from ChanAmpAssign
Using the popupmenu index for connecting headstages and amplifiers can
result in problems if e.g. the order of the amplifiers or their number
changes.

Therefore we now use the amplifier channel and serial number for
connecting both. This requires to increase the ChanAmpAssign wave version.

The additional call to HSU_UpdateChanAmpAssignPanel results in an up to
date GUI after locking the device.

Original-CommitID: 1939af568c43c9361c8c4296dddd14b8c4db3377