Remove the Testpulse from the list of available stim sets
Since cb550907 (Remove DA/TTL channel control adaptations for the
testpulse, 2015-10-29) we are already using a specific workflow for
the testpulse. Now we do the final transition and remove the testpulse
entry from the list of stimsets.

We also need to adapt all callers which assume that the first real
stimsets has index 2 (0-based). As that is now 1 (0-based).

We also don't need the function IsTestPulseSet anymore.

Introduce the function DAP_FormatStimSetPopupValue as we now can use the
same algorithm for DA and TTL channels.

Original-CommitID: 64d09b8d6b88455f440b6aeabc283aa31e7ce434