Amplifier <-> Headstage connection: Don't preselect NONE
Since f3a4da56 (Remove AmpIndex from ChanAmpAssign, 2016-03-21) we don't
use the index of the popupmenu anymore to recreate the connection between
the amplifier and the headstage.
The restoring idea was that in case the amplifier could not be found
(because the MCC app is not open) we choose NONE instead of a non-existing
amplifier. This approach, however, turned therefore silently the amplifier
off.

The new approach does not turn it off anymore, instead it warns the user,
and by keeping the currently selected amplifier does not confuse the user.

Original-CommitID: 0aa0b12d65ce3207febae0091f631d2c7109c605