Streamline labnotebook naming
The names of the four labnotebook wave are historically grown to be either
settingsHistory, numDocWave, numericValues or numericalValues (that is
only for the data wave of the numerical labnotebook).

This makes the code quite hard to read.

Therefore we now rename the labnotebook waves to:
- numericalValues
- numericalKeys
- textualValues
- textualKeys

For the AnalysisBrowser new wave getter functions were introduced for
handling the upgrade using UpgradeWaveLocationAndGetIt.

The sweep timestamp column wave is renamed to numericalValuesDat

Original-CommitID: fb83404116a77aa5f17ff85fbec5072941f3cd5a