Databrowser/Sweepbrowser: Add dedicated support for viewing dDAQ data
Data taken with "distributed DAQ" exhibits special features. Notably the
complete length of the data taken is a repetition of multiple stim sets.

Therefore we created a new viewing mode for dDAQ data.

This viewing mode has the following features:
- Don't display testpulse and data taken during the delays
- View each sweep wave slice, having the same length as the stimset, using
  a dedicated y axis
- Works with all overlay modes

For the sweep browser the assumption that every trace refers to a unique
wave is not true anymore, therefore we have to use UniqueWaveName on
duplicating the graph.

Original-CommitID: 69b4671990bc1b123488a480bd6e6248d839060e