CheckIfClose: Fix wrong results for two negative numbers
Prior to this commit the line
CheckIfClose(-0.8, -0.7)
returned 1 as the modulus was forgotten to be applied
to both numbers to be compared.

The bug is also in the original code in the unit-testing-framework's
function CLOSE_VAR. Which the author of this commit also authored.

Original-CommitID: a602dcd0cfc091e6146700963931384ceb1afe9f