CheckIfClose: Fix wrong results for two negative numbers
Prior to this commit the line
CheckIfClose(-0.8, -0.7)
returned 1 as the modulus was forgotten to be applied
to both numbers to be compared.

The bug is also in the original code in the unit-testing-framework's
function CLOSE_VAR. Which the author of this commit also authored.

Original-CommitID: f96119923a9dfb568b51eff9b6630c88f2454bd7