CheckIfClose: Fix wrong results for two negative numbers
Prior to this commit the line
CheckIfClose(-0.8, -0.7)
returned 1 as the modulus was forgotten to be applied
to both numbers to be compared.

The bug is also in the original code in the unit-testing-framework's
function CLOSE_VAR. Which the author of this commit also authored.

Original-CommitID: 2fa3e6dd6fa59af1069c9de7bbe02ca106735812