Bugfix: Don't overwrite pressure wave data on upgrade
In 3a1460aa (PressureControl: Allow the user to offset the applied
pressure, 2016-05-24) the numeric pressure wave version was upgraded from
1 to 2. Unfortunately the approach used therein did overwrite all user
settings.

We can not undo the overwriting from the switch 1->2 but we can prevent
happening it in the future on the upgrade by raising the version to 3.

Original-CommitID: f726aead5170baa384ede57bd3f0b7307c69f70a