DM_SaveAndScaleITCData: Enhance ITCDataWave saving
The current saving algorithm for the ITCDataWave was written prior to
07189f3a (Display the scaled AD data also during DAQ, 2016-01-18)
where we stopped displaying the ITCDataWave in the oscilloscope window.
This now makes it possible to leave the ITCDataWave unchanged and
streamline operations for speed.

Especially the changing of ITCDataWave's type is a big performance hit as
we have to change its type back on the next acquisition.

Now we employ the following strategy:
- Do the DA/AD scaling in the same wave assignment than copying the
  ITCDataWave
- Leave ITCDataWave unchanged

The user visible changes are as follows:
- The analysis functions now always receive the unscaled ITCDataWave.
  Before, the events Post Sweep, PostSet and Post DAQ got the scaled and
  floating point version. Users relying on the old behaviour should use
  AFH_GetLastSweepWaveAcquired instead.

Original-CommitID: fdbc7861b4b28bfe2d5cf85ae9f476e3314c468e