Oscilloscope: Fix graph updating logic in corner case
For TP MD during repeated acquisition the default axis ranges for the AD
graph are -1 to 1. This results in a very low spacing value and therefore
the graph is glued to the top and bottom of the graph window.

Work around that issue by using the data range for that case too.

Original-CommitID: 092b87f9b5600c1ba77e68337399d75f6db6e69e