IPNWB#WriteSingleChannel: Change channelSuffix handling
Previously you could use a channel suffix like "_1" and that got appended
like it is to the dataset name.

This makes it quite hard to use the channelSuffix for different purposes.

Now we let the caller add the "_" in front of the channelSuffix.

Original-CommitID: 7d68eaf110e7515bb71edc66d08e5140de302a08