GetLBTextualValues: Initialize wave
The labnotebook (textual and numerical) are resized on demand everytime
they are written into. Into the textual labnotebook is only written into
for DAQ and not for TP.

This leads to the fact that experiments without any data acquired, only
TP, can not be loaded into the analysis browser, as the number of columns
in the textual values and key wave differs.

Fix that by creating a textual labnotebook wave which has initially the
correct size and dimension labels.

Original-CommitID: e26b5a0641d07fbecac0516371f8eef4d656916f