Add HDF5 Browser ipf from IP7
Although our current one still works, it is not wise to use the version
which shipped with IP6 if there is a newer, but incompatible with IP6,
one.

Existing users have to change the shortcut from HDF to HDF-IP7 if using
Igor Pro 7.

Original-CommitID: cde51fe67b534807a6197534f0b7936ba4e44f99