NWB: Allow documenting the physical electrode
Most experiments have a constant mapping between the physical connection to the
cell (electrode) and the logical entity in MIES (headstage).

For more involved experiments it might be nice to be able to change that
mapping at runtime. The wave cellElectrodeNames which can be queried with
GetCellElectrodeNames(panelTitle) can be used for that.

Original-CommitID: e8c535b5f6fb2d8ba9c6f2cd8a156748a37eaef3