Labnotebook: Textual version, initialize index
In b9161ced (Rework labnotebook numeric and text code, 2015-03-20) we
added support for using NOTE_INDEX. But we forgot to set the initial index
for the textual values wave to 0.

This bug results in wasted memory only, as the first MINIMUM_WAVE_SIZE rows
are thought to be used already.

Original-CommitID: 8889bee554f78240884f9304483b603fc0ee1f93