AnalysisBrowser: Handle multiple experiments with the same name properly
When multiple experiments with the same name are encounterd we overwrote
the previous entries, as we derived the experiment folder from the name
only.

Now we derive the experiment folder from relative path to the
baseFolder. Additionally we also use the relative path for the displayed
experiment name to give the user a better idea on what data is shown.

Bug present since forever.

Original-CommitID: c14a6692367ed4e04bbcee8fd6559bc6d1fff6a1