Add EVIL_KITTEN_EATING_MODE
This mode allows to bypass any safety checks and aborts related to
hardware interaction. It therefore allows a developer to use MIES as far
as possible without having the appropriate hardware.

If this mode is turned on you are on your own!

Original-CommitID: 62dbb05765abc9f82c02d43ea371ddab94ffc8a7