Add support displaying textual labnotebook data
In the graph we show fake data, but attach to each point a tag with the
value of the labnotebook key. Empty entries are skipped and the indizes
are colored as done for the legend.

Currently no support for auto updating for newly acquired sweeps is done.

Original-CommitID: 473dac368b33938430410862a463fbd76b13b1ad