CreateTiledChannelGraph: Use zero as default value
We are using dDAQEnabled and oodDAQEnabled as boolean variables so we need
them to be not NaN. As NaN has a non consistent behaviour regarding if and
the ternary opertator.

Bug introduced in a116cca7 (CreateTiledChannelGraph: Add support for
oodDAQ data, 2016-08-10).

Original-CommitID: e88f743e827e6c75ba2d0a3eb90969ac33ecda67