Fix auto pipette offset buttons
The commit 04f295c0 (Auto Pipette Offset: Read the pipette offset for the
opposite clamp mode too, 2016-04-29) relied on the fact that the user in
the the correct clamp mode. This means the situation where the clamp mode
tab does not match the currently selected clamp mode is not handled.

Handle that situation as well by generalizing and merging the code for
both buttons and only rely on the currently active clamp mode.

Original-CommitID: b58207da96364c1eaba2f23d5baa950dcf393ad2