H5_WriteTextAttribute: Force dataspace to SIMPLE for lists
The NWB specification requires some attributes to have a dimension space
of type SIMPLE even if only one entry is present.

Use the new /A option to force a SIMPLE dataspace if a list is passed.
As this works only with IP7 a warning message is added for IP6.

Original-CommitID: eaa5e724d7556d6fc83150c0f7f49e8206b0c781