ZeroMQ Message Handler: Autostart on Igor Start
For this to work MIES_Include.ipf has to be located in "Igor Procedures"
and *not* in "User Procedures". See also the documentation to
IgorStartOrNewHook.

The #if exists acrobatic is for installation where the ZeroMQ XOP is not
available (e.g. our CI linux box).

Original-CommitID: 9eee1ff4b37bf7c0f893c7b791f38e8b26c786b5