Bugfix: Check for mismatched clamp mode earlier
Since 81daa711 (Propagate amplifier settings before DAQ/TP, 2016-05-11) we
query the gains and units from the MCC app. But this fails with an
assertion if the clamp modes in MIES and the MCC app don't match.

Therefore it is much nicer for the user if we do the "matching clamp mode
check" before the assertion could be triggered.

Original-CommitID: 4007c45cda6f469f7a03ba9fce4c929088cff2d5