Fix wrong formula generation in edge case
If we have lots of stimsets the shorthands will have a numeric suffix.
See WB_GenerateUniqueLabel.

In formulas like
B1/100
the current design leads to the replacement of B with the B stimset, and
not B1 with the B1 stimset as one would expect.

Iterate the epochCombineList from bottom to top so that we first try to
replace the shorthands with suffix.

Original-CommitID: 733ad5cf88505c737f9aa8fade8d8a76b693cf3b