fix: limited names for stimsets
bug 1: when a DA wave was named with some pattern containing TTL it was
assumed to be a TTL type on loading. example name: <noTTLwave>

bug 2: the return value for the DA/TTL channel type from the regex parsing function
was set the wrong way.

fix: use the regex pattern already in the beginning of the function and
ASSERT if the pattern was not found or differs

Original-CommitID: 9617bd019c65c7bc4399e80f60485539f6d42d34