Bugfix: Manual pressure control
We set the manual pressure multiple times if "All" is checked. This leads
to bugs as the order of the headstages matter, is grossly inefficent and
not what we want.

Now we use the user selected headstage pressure if all is selected and
manual pressure is active for this headstage. Otherwise we use
the headstage's pressure. Checking "all" and applying also overwrites the old manual
pressure values for each headstage with the one from the currently
selected user headstage.

Original-CommitID: 17af7af32e799ff9447f09aad68eae6a2de5b9ef