Bugfix: Disambiguate labnotebook entry search
In our labnotebook we store settings and data which were active during
DAQ, TP and possible other events.  We are querying the labnoteook by the
sweep number and search for the latest entry.

Let's assume now that the user has used different settings for a DAQ
and a manual TP after the sweep data was saved. Querying now
via GetLastSetting(numericalValues, 29, "DAC") the active DACs will return
the active DACs which were active during the TP and not during the DAQ.
Functions like NWB_AppendSweepLowLevel break due to that if the TP
settings were too different compared to the DAQ settings.
This means we have pxps which we can not convert to NWB.

The solution is two fold:
1.) Add a new labnotebook column "EntrySourceType" which documents for
    newly created labnotebooks the type of the source. It can be either DAQ/TP
    or unknown. Extend the interface of all labnotebook writing functions
    for setting that and all labnotebook querying functions
    passing that in.
2.) Add a heuristic which tries to identify the TP/DAQ blocks for existing
    labnotebooks. This heuristic is more involved in GetLastSetting
    compared to the one in GetLastSettingText. An approach of applying
    this heuristic once in filling in the "EntrySourceType" columns is not
    possible as we don't want to change the labnotebooks we read in from the
    analysisbrowser for these kind of things. Additionally if this
    heuristic needs further enhancement, we can do that much easier in that
    way.

Original-CommitID: d7bcb34c6fc52e519215359c141a96ee5f1ea16d