Pressure Devices: Make TTL B optional
For some experimental setups the required TTL B is not needed but must
still be provided by the hardware.

Add "-none-" to the list of available TTL A/B channels. To ease internal
handling the "-none-" entry is only selectable for TTL B.

Original-CommitID: 2d0eddb455df521521326214a0ee52891911dae5