HW_WriteDAC: Fix used units
The ITC XOP2 and its documentation disagree about the default state for
the V flag. And as it turns out, V=1 (passed value in volts and not in DAC
units) is not the default in the XOP but only in the documentation.

So we need to set it. Additionally we also set /C=1 in order to be
consistent with ITCReadADC2.

Original-CommitID: 27e754a9e5e3aa8ddc707bc89654d16d75765aed