AB_LoadDataWrapper: Fix return value
The function documentation states that the number of loaded objects *of the
given type" are returned.

But LoadData returns the number of loaded objects. And to that number all
subfolders are added on top.

As we need the number of loaded objects accurately we use GetListOfObjects
to count the just loaded objects.

The bug manifests itself on loading experiments which have subfolders in
root:MIES:ITCDevices:ITC1600:Device0:Data: as AB_LoadSweepConfigData does
in that case never finish.

Original-CommitID: e3f2904987e412f57c0174ba5c06866066bd04f5