ED_TPDocumentation: Avoid erroring out on old TP_Delta data
If you really really quickly start and stop the testpulse the function
TP_Delta, which updates BaselineSSAvg and friends, could not run, thus the
number of columns of these waves and the number of active headstages do
not coincide anymore.

For this case we just to nothing.

Original-CommitID: a9aa5dc43c566d08949f86022e0d6b93a7c7ec65