GetTPStorage: Fix wave note formatting on upgrade
In 6c5f16a0 (Generalize the wave note writing and reading functions used
in the TPStorageWave, 2014-09-23) we introduced common functions for
setting and getting wave note numbers. And the author of this commit
totally did not use it for GetTPStorage. Let's do that now.

The symptom of this bug is that some entries are duplicated in the wave
note on wave upgrades. And as ReplaceNumberByKey and friends only use the
first left most match this could have resulted in problems.

Original-CommitID: c245b826a97fe4daaa410a80b4eb1aee778796a3