WaveBuilder/Pulse Train: Make poisson distributed pulses reproducible
We use the same approach initially added in d53fc705 (Bugfix: Initialize
random seed for reproducible stim set creation, 2015-07-12).

And we also add a new "New Seed" button and a "Seed  / Sweep" checkbox to
the Pulse Train epoch.

The other todos (no overlap, fixed number of pulses) are bogus as it is
the nature of poisson distributed data that we should not fiddle with it.

The expression ln(abs(enoise(...))) models the inter-arrival times for a
poisson process.

As the epoch has not been reproducible before we can now make the switch
to the mersenne-twister PRNG.

Original-CommitID: c1318ffd32fb4583318a2b17cd46f86f0acf5c70