DM_StoreITCDataWaveScaled: Prevent storing invalid data
One report from last week stated that the user somehow created a sweep and
a config wave with differing channel number.

As we can not work with these corrupted data, we must ensure as early as
possible that the data is correct.

Original-CommitID: 9f35ed01adaf689a5fc48a57993035f83973f1ac