PGC_SetAndActivateControl: Set popStr for PopupMenues if not supplied
In the current version we did not pass the popStr to the PopupMenu
procedure if it was not supplied to PGC_SetAndActivateControl.

The reasons were that it is quite involved to find out the string from the
PopupMenu recreation macro.

Now we are exactly doing that.

Original-CommitID: c0e90f0093200278ed9a9483a4bb66be82e7a545