Add support for stopping and restarting DAQ on stimset change
Stopping DAQ, changing the stimset and then restarting DAQ is a common
pattern when doing experiments.

Currently this involves quite a few mouse clicks.

Make that easier by stopping the DAQ on the selected channel if that
channel is active. DAQ is also restarted if indexing is turned on and the
indexing end stimset is changed.

Original-CommitID: 0e51a6222f0e93a02adfa2e60e50892fe44479c4