AB_LoadDataWrapper: Fix empty folder removal
In 077d9d20 (AB_LoadDataWrapper: Cleanup after LoadData, 2015-07-03) we
introduced empty folder removal.

Unfortunately, as it turned out with the introduction of rtFunctionErrors
in 66d64ce1 (Add rtFunctionErrors pragma, 2017-01-04), the code did not
work as intended.

The indizes for GetIndexedObjNameDFR are not stable if entries are
removed from the same folder. And with rtFunctionErrors turned on Igor now
complains about invalid indizes passed to GetIndexedObjNameDFR.

The solution is as usual to iterate from the end.

See also the discussion in http://www.igorexchange.com/node/7557.

Original-CommitID: fa8d5a134b0be5927146be78edee11333b00dc0e