Upgrade SegWvType
dimensionlabel is used to get the total number of epochs in a stimset.
while at the save process SegWvType was updated to the newest version,
this is not the case when loading a wave.

Therefore the currently loaded SegWvType has to get upgraded to get correct dimension
labels.

Original-CommitID: 655966a47cb56ed114382105c2da9ee26118fe26