updated command line output
as stimsets are now loaded in a for loop the user is not aware which
stimset load has produced the error. Therfore we add the name of the
stimset as an optional parameter to the Loading Function and update the
printf statements.

Original-CommitID: 74e165e4e2574865183a6f5514827815b623865c