CreateTiledChannelGraph: Enhance axis positioning in dDAQ mode
Altough we hide the axes in dDAQ mode the user can still change the axis
scaling by using the mouse wheel when he/she is over the axis. But for
that to work the axes positions must be at a place where they are
expected.

Move them below and left to the data plots.

Additionally we lower the bottom margin to use the space as good as
before.

Original-CommitID: 48959c440ad27a583086225d9ea35970918ab504