EquallySpaceAxis: Allow to have some axes at the very front and very back
In the old implementation EquallySpaceAxis always either equally spaced
the axes in the order which AxisList returned or in a given sortOrder.

In some cases users want to have a list of axes always at the
top/bottom of the graph, this can now be done with the optional
listForBegin/listForEnd parameters.

Original-CommitID: fcf0afe886efbcb941bc4b389b0764d951807420